[#11258] YCQL: Add timeout for prepare child
Summary:
We will first do a partial solution for the deadlock of threads during the PrepareChild
method. The partial solution is to add a timeout for the prepare child transaction, in which we will
time out if no response is received before then.

The full solution (making PrepareChild fully async) is under development, but this is the fallback
if we decide that is more complex than necessary.

Test Plan: ybd --cxx-test cql-index-test --gtest_filter=CqlIndexTest.TestSaturatedWorkers

Reviewers: bogdan, sergei, pjain

Reviewed By: sergei, pjain

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D15254