[#6428] DST: Ensure uniform distribution of tablets right at create table time
Summary:
Currently due to an old Kudu artifact, when we assign tablet replica to tserver during CreateTable
we choose a tserver randomly using reservoir sampling and put the replica there. But this does not
yield the most uniform placement as has been seen before. The LB is then responsible for
distributing and balancing it. There are several consequences like the number of RBS increases. It
also make backup/restore more error-prone as during restore we create tables and expect their
placement not change after create and throughout the period of restore.

This diff borrows logic from the Load Balancer and introduces similar logic to assign replicas based
on per table and global load distribution of the tservers. Note that it is meant to be a short-term
fix and in the long-term we should have a way to unify both the code pieces (LB and CM).

Test Plan: ybd --cxx_test create-table-itest --gtest-filter CreateTableITest.TestSpreadReplicasEvenly

Reviewers: rahuldesirazu, bogdan, jhe

Reviewed By: jhe

Subscribers: kannan, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D15349