[#13211] YSQL: Integrate READ COMMITTED isolation with wait queues
Summary:
READ COMMITTED isolation provides blocking i.e., waiting semantics during
transaction conflicts by retrying a query indefinitely on kConflict errors with
exponential backoff. Once all conflicting transactions end, the next retry of
the query will successfully run. The exponential backoff delay between retries
is to ensure that the YSQL backend doesn't overwhelm the system by retrying
in a tight loop.

D17304 (dc811063af47a1727cbe59e5ce621f2e94da71c9) added the new wait
queue based implementation which can be used if the tserver gflag
enable_wait_queues is set to true. In this case, a read/ write rpc from
a YSQL backend to the transaction participant is blocked if there are conflicts
detected on the participant and unblocked once all conflicting transactions have
completed (either committed or aborted). There are 2 scenarios possible once
the rpc is unblocked -

(1) It is still conflicting because some transaction has committed which made a
conflicting modification to the data.
(2) No transaction with a conflicting modification to the data has committed.

For case (1), the kConflict error is returned to the YSQL backend. In (2), the
rpc makes progress and returns the appropriate result to the YQSL backend. The
waiting is transparent to the YSQL backend (i.e., not differentiable from a rpc
which wasn't blocked).

So, if enable_wait_queues is set, there is no need for a READ COMMITTED
transaction to sleep before retrying the query (because a kConflict error, if at
all [with case (1)] is sent only after all conflicting transactions have ended).

NOTE: REPEATABLE READ and SERIALIZABLE isolation levels also retry with
exponential backoff when kConflict errors occur in the first statement of a
transaction. This too is not needed if enable_wait_queues is true. This diff
ensures that as well.

Other miscellaneous changes -

(1) SKIP LOCKED wasn't working for single shard transactions earlier, fixed it.

(2) Changed naming at most places except docs from "pessimistic" and "optimistic"
locking to "Wait-on-Conflict" and "Fail-on-Conflict". This is an effort to move away
from the words "pessimistic" and "optimistic" since they are wrongly defined and
used in internal discussions and their meanings cause confusion in external
discussion. For example, when we say optimistic, people think we mean
"optimistic concurrency control" which is widely known in literature and
industry (like here - https://people.eecs.berkeley.edu/~fox/summaries/database/optimistic_concurrency.html)

Created GitHub issue #14935 to situations where nodes can have different values for
enable_wait_queues during the rolling restart.

Test Plan:
./yb_build.sh --java-test org.yb.pgsql.TestPgIsolationRegress#isolationRegressWithWaitQueues
./yb_build.sh --java-test org.yb.pgsql.TestPgIsolationRegress#withDelayedTxnApplyWithWaitQueues

Reviewers: tvesely, sergei, rsami

Reviewed By: rsami

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D20974