[PLAT-5530] Rename "Full Name" to "Customer Name"
Summary:
**Context**
JIRA: https://yugabyte.atlassian.net/browse/PLAT-5530
When viewing the `user profile` we are presented with the following:
{F31473}
This "full name" is not the name of the user. It is actually just the customer name.
Furthermore, we don't request or store any information about the user's name.

Thus, the label “Full Name” shown in the UI is misleading and should be renamed.

**Change**
Change the UI label for the name field from "Full Name" to "Customer Name".

Test Plan:
Verify the label text has changed on the user profile page.
{F31474}
{F31475}

Reviewers: sb-yb, shagarwal

Reviewed By: shagarwal

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D20481