Fixing flaky AccessManagerTest
Summary:
Hamid reported that the AccessManagerTest#testCreateKubernetesConfigFileExists test has been flaky.
Found that this test had a bug and has been silently passing all along. The path to place the
kubeconfig file is expected to be `/tmp/yugaware_tests/amt/keys/` but was set to
`/tmp/yugaware_tests/keys/`. The test was always creating a new kubeconfig file and not doing the
intended test.

I have added an explicit `fail()` if a new kubeconfig is created. Also fixed the path so that this
ut actually tests that createKubernetesConfig() fails with the intended exception.

Test Plan: Validated the unit test passes.

Reviewers: gjalla, hzare

Reviewed By: gjalla, hzare

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D20554