[#14802] [TEST] ysql: yb_pg_errors fails in master branch
Summary:
After commit 6875e1b, the test fails consistently.

This is one of the different hunks in test output:
```
*** ${TEST_TMPDIR}/pgregress_output/yb_pg_misc_serial_schedule/expected/yb_pg_errors.out    2022-11-02 04:43:48.490157772 +0000
--- ${TEST_TMPDIR}/pgregress_output/yb_pg_misc_serial_schedule/results/yb_pg_errors.out 2022-11-02 04:43:48.430153470 +0000
***************
*** 564,570 ****
  INSERT INTO test_temp_tab VALUES (4) RETURNING xmin;
   xmin
  ------
!     7
  (1 row)

  INSERT INTO test_temp_tab VALUES (5) RETURNING xmax;
--- 564,570 ----
  INSERT INTO test_temp_tab VALUES (4) RETURNING xmin;
   xmin
  ------
!    13
  (1 row)

  INSERT INTO test_temp_tab VALUES (5) RETURNING xmax;
```
This was due to addition of temp tables to yb_pg_copy which is above yb_pg_errors test.
If I comment out the test for yb_pg_copy, the values shown in the test output of yb_pg_errors would be different.

Test Plan: ybd --java-test 'org.yb.pgsql.TestPgRegressPgMisc'

Reviewers: dsrinivasan

Reviewed By: dsrinivasan

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D20720