[PLAT-5890] Fixing logic to check Pod Readiness
Summary:
In some kubernetes universes we were marking the pod as Ready when it
wasn't ready to recieve traffic. This caused us to incorrectly assume pod
is Ready when it is marked for deletion.
This change reworks the readiness check to see if the pod is running
or not.

This change also refactors some of the code alongside to introduce a
getPodObject method on KubernetesManager class, since the existing
getPodStatus method is not enough to make this decision.

Test Plan:
Ran itest,
Fixed broken unit test.
Verified by deploying a universe locally in GKE on the local dev Build of platform.
Restarting pods in a universe and verified that it completed.
{F32957}

Reviewers: nbhatia, nsingh, bgandhi, sneelakantan, sanketh

Reviewed By: bgandhi, sneelakantan, sanketh

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D20640