[PLAT-6160]: Fix the VM Image upgrade issue with YBC.
Summary:
Added the changes to fix the VM image upgrade issue with ybc.
Also, Added the change to inherit ybc-based parameters directly from the universe details so that we don't have to rely on the payload.

UI changes to only send the enableYBc as true for software upgrades and for other tasks YBA will inherit from universe details.

Test Plan: Tested manually by creating the universe and successfully upgrading the VM image (on both ybc and non-ybc universe) (keeping both enableYbc as true and false)

Reviewers: asathyan, vpatibandla, kkannan, yshchetinin

Reviewed By: vpatibandla, kkannan, yshchetinin

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D20882