[#13750] Add DEFINE_RUNTIME_uint32, DEFINE_NON_RUNTIME_uint32
Summary:
Commit e649ffd81edac569dbad713826593cd2d1d11c8b

    [#13750] Add explicit gflag macros for DEFINE_RUNTIME and DEFINE_NON_RUNTIME

adds DEFINE_RUNTIME_foo and DEFINE_NON_RUNTIME_foo macros for most
types, but it doesn't do so for the uint32 type.  There are currently no
uses of the uint32 type (likely because it was not supported until the
recent gflags thirdparty update), but there may be in the future, so
prepare for that case by adding the macros now.  The linter at
yugabyte/yb-cpplint can be updated soon after.

Test Plan: Jenkins: compile only

Reviewers: bogdan, hsunder

Reviewed By: hsunder

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D20910