[#14909] CDCSDK: Fix issue where 'cdcsdk_change_event_count' metric is incremented even on empty batches
Summary:
In every GetChanges call if the 'last_sent_checkpoint' is greater than the 'from_op_id' in the GetChanges call, we reset the cached schema details in cdc_service.

But there was a bug in the part where we updated the 'last_sent_checkpoint' at the end of GetChanges which meant :
The DDL record was sent in every GetChanges call, as a result of which 'cdcsdk_change_event_count' was incremented by an additional value.
Each GetChanges call repopulated the cached schema, resulting in a performance degradation

Fixing the line where 'last_sent_checkpoint' is updated.

Test Plan: Added ctest: TestCDCSDKChangeEventCountMetricUnchangedOnEmptyBatches

Reviewers: sdash, srangavajjula, skumar

Reviewed By: sdash, skumar

Differential Revision: https://phabricator.dev.yugabyte.com/D20931