[PLAT-6266][Platform] Using universeDefinitionTaskParams struct for node tasks instead of default values of NodeTaskParams.
Summary:
During merging we are overriding some values which we are not supposed to.
While adding a node to universe, we are taking merging NodeTaskParams() into UniverseDetailsTaskParams() but default nodetaskparams had rootAndClientRootCASame set to true but universe in DB had false which would be overriden during merge.

Test Plan: Ran failing itest - testcloudnodeops-cloud-config-aws-custom-ami-paid.

Reviewers: sanketh, shagarwal, yshchetinin

Reviewed By: shagarwal, yshchetinin

Subscribers: shagarwal, jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D21075