[#14891] docdb: Cleanup load balancer tests
Summary:
We had some incorrect lines using ASSERT_OK(IsLoadBalanced()) in tests. We were passing in an
incorrect value to this function which was causing it to return an error, but since it returns a
result<bool> and we were using ASSERT_OK instead of ASSERT_RESULT, this was being ignored.

Fixing this in a few spots, but also removing the check in other spots as we already perform the
is_idle check that IsLoadBalanced checks for in WaitForLoadBalanceCompletion.

Also cleaning up some comments and code around the lb colocated test.

Test Plan:
```
ybd --cxx-test load_balancer_multi_table-test
ybd --cxx-test load_balancer_colocated_tables-test
```

Reviewers: skedia, asrivastava

Reviewed By: asrivastava

Subscribers: sshah, ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D20906