[#15009] CDCSDK: Fix for the flaky test CDCSDKYsqlTest.TestCDCSDKMetricsTwoTablesSingleStream
Summary: This diff fixes the flakiness of CDCSDKYsqlTest.TestCDCSDKMetricsTwoTablesSingleStream. The flakiness was due to a flag cdc_state_checkpoint_update_interval_ms which was set to zero, because of which in some cases getChanges call was trying to retrieve intents whose "Apply" record's OpId was lesser than the checkpoint.

Test Plan:
Jenkins: test regex: .*TestCDCSDKMetrics.*
Ran the test multiple times on dev server

Reviewers: skumar, vkushwaha, rvenkatesh, sdash, abharadwaj

Reviewed By: sdash, abharadwaj

Subscribers: ycdcxcluster

Differential Revision: https://phabricator.dev.yugabyte.com/D21143