[#15032] CDCSDK: Alter table support with CDC snapshot
Summary:
Consider a scenario where CDC is taking a snapshot for //test_1// table, which has currently 3 columns. there is a possibility that in the past same table was altered multiple times. so when taking a snapshot a table, we should pick the data based on the current schema.

As part of this fix, during snapshot processing, CDC will collect the latest schema from the system catalog and will keep it in its local cache, so that the same cache can be referred to get the table schema for the next snapshot batch records decoding.

Test Plan:
Jenkins: skip
./yb_build.sh --cxx-test integration-tests_cdcsdk_ysql-test --gtest_filter CDCSDKYsqlTest.TestMultipleTableAlterWithSnapshot

Reviewers: skumar, srangavajjula, abharadwaj

Reviewed By: abharadwaj

Subscribers: ycdcxcluster

Differential Revision: https://phabricator.dev.yugabyte.com/D21144