[#15044] YSQL: Fix initial_sys_catalog_snapshot dependencies
Summary:
Internally the scripts implementing the initial_sys_catalog_snapshot
build target use run-with-timeout tool, which is considered a part of
the test infrastructure and is not build if --no-tests is specified.

Add run-with-timeout to the initial_sys_catalog_snapshot's dependency
list to make sure it is built.

Test Plan:
ybd --no-tests reinitdb
Not a product issue, does not require automated tests

Reviewers: mbautin, steve.varnau

Reviewed By: steve.varnau

Subscribers: yql

Differential Revision: https://phabricator.dev.yugabyte.com/D21165