[#13733] docdb: Differentiate ybctid and dockey usage for colocation
Summary: Currently dockeys and ybctids are used interchangeably in our code, which can be a bit confusing as there are some key differences. Notably, the main difference is that ybctids do not store any info about colocation ids or cotable ids. As part of simplifying colocation usage across ysql, this diff introduces a new `PgsqlYbctid` class that is just used for ybctids, and that has no references to colocation. Also introduces a `DocKeyBase` abstract class that both `PgsqlYbctid` and `DocKey` inherit from.

Test Plan: Jenkins

Reviewers: alex

Reviewed By: alex

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D19717