[#15063]CDCSDK: Alter table support is broken with packed row
Summary: During the unit test case failure analysis we observed that for packed row decoding CDC need the corresponding schema and schema version in the //SchemaPackingStorage// object so that in the later stage it can decode the packed row in //PopulatePackedRows// method. but in the method //PopulateCDCSDKWriteRecord//, it is always taking the latest schema version, causing this issue. to handle this we will then pass the corresponding schema version, that CDC is collected from the system catalog table to //PopulateCDCSDKWriteRecord// method

Test Plan: Running all the alter table unit testcases with FLAGS_ysql_enable_packed_row as true

Reviewers: skumar, srangavajjula, abharadwaj

Reviewed By: abharadwaj

Subscribers: ycdcxcluster

Differential Revision: https://phabricator.dev.yugabyte.com/D21246