[#15104] DocDB, Colocation: Fixes primary table being renamed on a colocated table rename
Summary: On renaming a colocated table, the parent table was getting renamed in the superblock instead. It was happening because of using the default table id (primary table id) while setting the name. This diff fixes the bug.

Test Plan:
Manually verfied that the table name is correct in the superblock. On renaming a colocated table `t1` to `new_name`:

Tables in superblock state without the fix:
```
    table_id: "000033e8000030008000000000004000.tablegroup.parent.uuid"
    table_name: "new_name"

    table_id: "000033e8000030008000000000004001"
    table_name: "t1"
```

Tables in superblock with fix:
```
    table_id: "000033e8000030008000000000004000.tablegroup.parent.uuid"
    table_name: "000033e8000030008000000000004000.tablegroup.parent.tablename"

    table_id: "000033e8000030008000000000004001"
    table_name: "new_name"
```

Reviewers: skedia, alex

Reviewed By: alex

Subscribers: ybase, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D21285