[#15062]CDCSDK:Differentiate between Update and Insert records in case of packed row
Summary:
In CDC, we decode intents to form records, and we have made an assumption that if an intent's type is packed column// (docdb::ValueEntryType::kPackedRow),// it came from an insert operation. But some of the test cases failing are because we are updating all non-key columns of a table with packed row enabled, and we wrongly assume it is an insert operation. We essentially need to distinguish between intent's belonging to an update operation (all non-key columns updated) vs intents belonging to an insert operation, in the case of packed rows.

We did an //sst// file analysis for packed vs unpacked using //sst-dump//, we have observed if all the columns of a record are updated, it comes as single entry in the //sst// file for the packed row whereas for an unpacked row, there are separate entries for all the updated columns. Here is a snapshot for that:-

**Packed Row entire row update **
Sst file format: block-based
Not found (yb/docdb/schema_packing.cc:173): Schema packing not found: 0: . Key: SubDocKey(DocKey(0xc0c4, [2], []), [HT{ physical: 1669187931133661 }])
[sureshdash_yb@dev-server-yb-sureshdash-2 bin]$
s
**Unpacked row  entire row update**
Sst file format: block-based
SubDocKey(DocKey(0x1210, [1], []), [ColumnId(11); HT{ physical: 1669190580751924 }]) -> 130
SubDocKey(DocKey(0x1210, [1], []), [ColumnId(12); HT{ physical: 1669190580751924 w: 1 }]) -> 140
SubDocKey(DocKey(0x1210, [1], []), [ColumnId(13); HT{ physical: 1669190580751924 w: 2 }]) -> 150

Because there is no way to differentiate INSERT VS UPDATE operation(when all columns are updated) in a packed row, so we are planning to modify our test cases to verify the excepted count based on the condition. As there is no data loss, we will handle it from the test case it self.

Test Plan:
Jenkins: skip
skipping detective run as a previous run detective looks good, want to rebase with master and land it.
Running All the failure testcases with packed row.

Reviewers: skumar, srangavajjula, abharadwaj

Reviewed By: abharadwaj

Subscribers: ycdcxcluster

Differential Revision: https://phabricator.dev.yugabyte.com/D21349