[#15156] YSQL: Have joinrel param batching info acknowledge that batched clauses should not be imposed on child rels that can't handle batching
Summary:
Before this change, it was possible to have a series of batched joins of the form
```
Batched NL J1
-> Seq scan on s2
-> Batched NL J2
Join Filter: ((s2.r3 = s1.r3) AND (s1.r1 = s3.r1))
------> Seq scan on s1
------> Index only scan on s3 (Index Cond: ((r1 = ANY (ARRAY[s1.r1, $1, $2])) AND (r2 = ANY (ARRAY[s1.r2, $3, $4])))
```
This is problematic as it allows batches of s2 to be passed down to J2. s1 here is a sequential scan that is not ready to accept batches of outer tuples as the index scan of s3 is. Hence, the join filter `s2.r3 = s1.r3` makes no sense as for each iteration of J2 we have a batch of s2 tuples but only one tuple for s1.

This change alleviates this issue by making the param_info of a joinrel check to see if the outer rel of the join is involved with any batchable joinrestrictclauses that require other parameterized relations that are not in the outer path's param_info->yb_ppi_req_outer_batched.

Through testing this case, it appeared that the BNL hashtable was not using the same expression context as its owning nestloopnode, causing it to crash when evaluating parameter values. This is also fixed in this change.

Test Plan: ./yb_build.sh release --java-test 'org.yb.pgsql.TestPgRegressJoin'

Reviewers: amartsinchyk, mtakahara

Reviewed By: mtakahara

Subscribers: mihnea, smishra, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D21380