[PLAT-6377] Seting instance type in node details of universe
Summary:
When changing the instance type of the pods from the edit universe UI, the final step is to
update the universe details in the YBA db. At this step, the instanceType of the nodes should be
updated. This change copies over the instanceType from the user intent of this operation.

Test Plan:
Verified that the YBA db has the universe details correctly updated with new instance
type. Also verified that subsequent edit universe happens successfully now without the UI issue
reported in PLAT-6377.

Reviewers: gjalla, sanketh, bgandhi, anijhawan

Reviewed By: bgandhi, anijhawan

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D21403