[PLAT-5044][xCluster][K8s] Automatically update the source master address in the target universe cluster config for K8s universes
Summary:
This diff adds a subtask to the `editCluster` method to update the
source master addresses on the target universe cluster config if
they have changed.

Test Plan:
Ran the EditUniverse task on a K8s universe and made sure the
master addresses are updated on the target universe cluster config.

Reviewers: sanketh, gjalla, amalyshev

Reviewed By: amalyshev

Subscribers: jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D21367