[#14072] YSQL: Stabilize PgTableSizeTest.MaterializedViewSize
Summary:
The test was failing because the size of the materialized view
sometimes turned out to be 0 even though it had data in it. This is
because sometimes the query for the materialized view size was landing
before the TServer could calculate and report the drive size.

The above was happening because the test tries to speed up the drive
size reporting by reducing FLAGS_tserver_heartbeat_metrics_interval_ms
when the test begins. However this is not a runtime flag and needs to
be set as part of SetUp().

The failure was sporadic because the wait-time for flush and compact,
and then a subsequent sleep totals about 5 seconds which is the default
value of this flag.

Test Plan:
Jenkins: test regex: .*PgTableSizeTest.*

Without the changes, the test always failed at least once in Detective. The runs are now clean.

Reviewers: fizaa

Reviewed By: fizaa

Subscribers: yql, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D21469